# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestion PartList",
        "contact": {
            "email": "antony.rodriguez@mimco.com.pe"
        },
        "version": "1"
    },
    "paths": {
...
        "/GestionPartList/public/index.php/regi_proy_tarea": {
            "post": {
                "tags": [
                    "Gestion Proyecto Tarea"
                ],
                "summary": "Regitrar un proyecto tarea.",
                "operationId": "App\\Http\\Controllers\\ProyectotareaController::regi_proy_tarea",
                "parameters": [
                    {
                        "name": "intIdProyZona",
                        "in": "path",
                        "description": "Ingrese el id proyecto zona",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "intIdProy",
                        "in": "path",
                        "description": "Ingrese el id proyecto",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "intIdTipoProducto",
                        "in": "path",
                        "description": "Ingrese el id TipoProducto",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "varDescripTarea",
                        "in": "path",
                        "description": "Ingrese el DescripTarea",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "acti_usua",
                        "in": "path",
                        "description": "Ingrese el DescripTarea",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "intIdProyZona": {
                                        "type": "string"
                                    },
                                    "intIdProy": {
                                        "type": "string"
                                    },
                                    "intIdTipoProducto": {
                                        "type": "string"
                                    },
                                    "varDescripTarea": {
                                        "type": "string"
                                    },
                                    "acti_usua": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "intIdProyZona": "1200",
                                    "intIdProy": "1260",
                                    "intIdTipoProducto": "1410",
                                    "varDescripTarea": "19_0021_2"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sin Mensaje"
                    },
                    "407": {
                        "description": "El Documento de identidad ingresado no se encuentra registrado."
                    },
                    "404": {
                        "description": "No se encuentra el metodo"
                    }
                }
            }
        },
        "/GestionPartList/public/index.php/vali_tarea": {
            "post": {
                "tags": [
                    "Gestion Proyecto Tarea"
                ],
                "summary": "validar si la tarea existe",
                "operationId": "App\\Http\\Controllers\\ProyectotareaController::vali_tarea",
                "parameters": [
                    {
                        "name": "intIdProyZona",
                        "in": "path",
                        "description": "ingresar el id proyecto zona",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "intIdProy",
                        "in": "path",
                        "description": "ingresar el id proyecto",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "intIdTipoProducto",
                        "in": "path",
                        "description": "ingresar el id  tipo producto ",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "varDescripTarea",
                        "in": "path",
                        "description": "ingresar la descriptarea ",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "intIdProyZona": {
                                        "type": "string"
                                    },
                                    "intIdProy": {
                                        "type": "string"
                                    },
                                    "intIdTipoProducto": {
                                        "type": "string"
                                    },
                                    "varDescripTarea": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "intIdProyZona": "25214121",
                                    "intIdProy": "1500",
                                    "intIdTipoProducto": "1250",
                                    "varDescripTarea": "114"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "si no existe entonces, mandamos un mensaje de exito para la siguiente evaluacion"
                    },
                    "201": {
                        "description": " la tarea ya se encuentra asignada a una zona.Por favor verificar el Patlist"
                    },
                    "404": {
                        "description": "No se encuentra el metodo"
                    }
                }
            }
        },
...
    }
}